+++ src/DVarr/Data/Migrations/20260720105840_AddEventBroadcast.cs
@@ -1,0 +1,28 @@
+﻿using Microsoft.EntityFrameworkCore.Migrations;
+
+#nullable disable
+
+namespace DVarr.Data.Migrations
+{
+    /// <inheritdoc />
+    public partial class AddEventBroadcast : Migration
+    {
+        /// <inheritdoc />
+        protected override void Up(MigrationBuilder migrationBuilder)
+        {
+            migrationBuilder.AddColumn<string>(
+                name: "Broadcast",
+                table: "Events",
+                type: "TEXT",
+                nullable: true);
+        }
+
+        /// <inheritdoc />
+        protected override void Down(MigrationBuilder migrationBuilder)
+        {
+            migrationBuilder.DropColumn(
+                name: "Broadcast",
+                table: "Events");
+        }
+    }
+}